Move strippublic.h into JKA source tree.
