+++ code/qcommon/strippublic.h
